% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_halfeye_plot.R
\name{beta_halfeye_plot}
\alias{beta_halfeye_plot}
\title{Make a halfeye ggplot for all beta values from a `brms` model}
\usage{
beta_halfeye_plot(model, start_string = "b_")
}
\arguments{
\item{model:}{a `brms` model object}

\item{start_string:}{what is the starting string of the model parameters that should be plotted. Defaults to "b_" to get all fixed beta values.}
}
\value{
a halfeye ggplot object with one halfeye for each beta parameter
}
\description{
Make a halfeye ggplot for all beta values from a `brms` model
}
\examples{
m1 \%>\% beta_halfeye_plot()


}
