% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct_cols.R
\name{distinct_cols}
\alias{distinct_cols}
\title{Select only unique columns}
\usage{
distinct_cols(.data, keep = NA, reverse = FALSE)
}
\arguments{
\item{.data:}{a dataframe whose columns will be evaluated for uniqueness and then have duplicates removed}

\item{keep:}{a vector of quoted column names for columns that should be kept regardless of duplicate status. Defaults to NA}

\item{reverse:}{should columns be evaluated for duplicates in reverse order? Defaults to FALSE}
}
\value{
a dataframe with only unique columns
}
\description{
Select only unique columns
}
\examples{
mtcars2 <- mtcars \%>\% mutate(hp2 = hp)
mtcars2 \%>\% distinct_cols()

}
