% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghist.R
\name{gghist}
\alias{gghist}
\title{Quickly make a ggplot histogram from a dataframe and variable name}
\usage{
gghist(.data, var, bins = NULL)
}
\arguments{
\item{.data:}{a vector of values, such as one returned by `rnorm()`}

\item{bins:}{how many bins the histogram uses. Defaults to ggplot's default}
}
\value{
a histogram ggplot object
}
\description{
Quickly make a ggplot histogram from a dataframe and variable name
}
\examples{
mtcars \%>\% gghist(mpg)


}
