% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggskim.R
\name{ggskim}
\alias{ggskim}
\title{Create ggplots to quickly explore data}
\usage{
ggskim(.data, ...)
}
\arguments{
\item{.data:}{a dataframe to make plots from}
}
\value{
either a list of the plots or simply the plots themselves
}
\description{
Create ggplots to quickly explore data
}
\examples{
iris \%>\% ggskim()


}
