% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_aesthetics.R
\name{minimal_ggplot_theme}
\alias{minimal_ggplot_theme}
\title{A nice minimal ggplot theme}
\usage{
minimal_ggplot_theme(base_family = "Roboto Condensed", gridlines = F)
}
\arguments{
\item{base_family:}{the base font family, defaults to "Roboto Condensed"}

\item{gridlines:}{should gridlines be displayed? defaults to F}
}
\value{
A ggplot theme
}
\description{
A nice minimal ggplot theme
}
\examples{
mtcars \%>\% ggplot(aes(x = wt, y = mpg)) + geom_point() + minimal_ggplot_theme()
}
