% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_names.R
\name{nice_names}
\alias{nice_names}
\title{Make the names nice}
\usage{
nice_names(.data, ...)
}
\arguments{
\item{.data:}{a dataframe whose names are to be changed}
}
\value{
a dataframe with nice names
}
\description{
Make the names nice
}
\examples{
iris \%>\% nice_names()


}
