% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_breaks.R
\name{quantile_breaks}
\alias{quantile_breaks}
\title{Get a vector of quantile values from a vector}
\usage{
quantile_breaks(value, prob, digits = 1)
}
\arguments{
\item{prob:}{a sequence of probability values, ie c(0, 0.5, 1) or seq(0, 1, 0.25)}

\item{digits:}{number of digits to round the quantiles to}
}
\value{
a vector of quantile values rounded to a specified number of digits
}
\description{
Get a vector of quantile values from a vector
}
\examples{
quantile_breaks(1:10, c(0,0.5,1), 2)
}
