% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_trans.R
\name{quantile_trans}
\alias{quantile_trans}
\title{A transformation function to be used in ggplot2 scales}
\usage{
quantile_trans(val, prob, digits)
}
\arguments{
\item{prob:}{a sequence of probability values, ie c(0, 0.5, 1) or seq(0, 1, 0.25)}

\item{digits:}{number of digits to round the quantiles to}
}
\value{
a scale that is untransformed but has breaks specified by quantiles
}
\description{
A transformation function to be used in ggplot2 scales
}
