% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_x_quantile.R
\name{scale_x_quantile}
\alias{scale_x_quantile}
\title{Set the x axis scale to continuous with quantile breaks}
\usage{
scale_x_quantile(val, prob = seq(0, 1, 0.25), digits = 1, ...)
}
\arguments{
\item{prob:}{a sequence of probability values, ie c(0, 0.5, 1) or seq(0, 1, 0.25)}

\item{digits:}{number of digits to round the quantiles to}
}
\value{

}
\description{
Set the x axis scale to continuous with quantile breaks
}
