% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_y_quantile.R
\name{scale_y_quantile}
\alias{scale_y_quantile}
\title{Set the y axis scale to continuous with quantile breaks}
\usage{
scale_y_quantile(val, prob = seq(0, 1, 0.25), digits = 1, ...)
}
\arguments{
\item{val:}{the column to calculate breaks from, ie mtcars$mpg}

\item{prob:}{a sequence of probability values, ie c(0, 0.5, 1) or seq(0, 1, 0.25)}

\item{digits:}{number of digits to round the quantiles to}
}
\value{
a ggplot2 scale
}
\description{
Set the y axis scale to continuous with quantile breaks
}
