% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mcm.R
\name{theme_mcm}
\alias{theme_mcm}
\title{A light ggplot theme for MCM's personal use}
\usage{
theme_mcm(
  base_family = "FrauncesSuperSoftWonky-Light",
  title_family = "FrauncesSuperSoftWonky-BlackItalic",
  subtitle_family = "FrauncesSuperSoftWonky-LightItalic",
  axis_family = "FrauncesSuperSoftWonky-Light",
  base_color = "gray10",
  primary_color = "white",
  accent_color = "black",
  base_size = 10,
  gridlines = F,
  facet_outlines = F
)
}
\arguments{
\item{base_family:}{the base font family}

\item{title_family:}{font to be used for the plot title}

\item{subtitle_family:}{font to be used for the subtitle}

\item{axis_family:}{font to be used for the axis titles}

\item{base_color:}{base color for text and lines}

\item{primary_color:}{color for primary components}

\item{accent_color:}{color for accent components}

\item{base_size:}{base font size}

\item{gridlines:}{should gridlines be displayed?}

\item{facet_outlines:}{should facets be outlined?}
}
\value{
A ggplot theme
}
\description{
A light ggplot theme for MCM's personal use
}
\examples{
mtcars \%>\% ggplot(aes(x = wt, y = mpg)) + geom_point() + theme_mcm()
}
