% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mcm.R
\name{theme_mcm_dark}
\alias{theme_mcm_dark}
\title{A dark ggplot theme for MCM's personal use}
\usage{
theme_mcm_dark(
  base_family = "fraunces",
  title_family = "fraunces",
  subtitle_family = "fraunces",
  axis_family = "fraunces",
  base_color = "white",
  primary_color = "grey10",
  accent_color = "grey90",
  gridlines = F,
  facet_outlines = F,
  base_size = 10,
  mult1 = 1.2,
  mult2 = 1.4,
  mult3 = 1.5,
  mult4 = 2.25,
  large_lineheight = F
)
}
\arguments{
\item{base_family:}{the base font family}

\item{title_family:}{font to be used for the plot title}

\item{subtitle_family:}{font to be used for the subtitle}

\item{axis_family:}{font to be used for the axis titles}

\item{base_color:}{base color for text and lines}

\item{primary_color:}{color for primary components, including background}

\item{accent_color:}{color for accent components}

\item{base_size:}{base font size}

\item{gridlines:}{should gridlines be displayed?}

\item{facet_outlines:}{should facets be outlined?}

\item{large_lineheight:}{should the subtitle lineheight be increased to 1 to prevent overlaps?}
}
\value{
A ggplot theme
}
\description{
A dark ggplot theme for MCM's personal use
}
\examples{
mtcars \%>\% ggplot(aes(x = wt, y = mpg)) + geom_point() + theme_mcm_dark()
}
