% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggle_error.R
\name{toggle_error}
\alias{toggle_error}
\title{Toggle error handling between rlang::entrace() and recover()}
\usage{
toggle_error()
}
\description{
Toggle error handling between rlang::entrace() and recover()
}
