% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_hist.R
\name{vector_hist}
\alias{vector_hist}
\title{Quickly make a ggplot histogram from a single vector}
\usage{
vector_hist(.data, bins = NULL)
}
\arguments{
\item{.data:}{a vector of values, such as one returned by `rnorm()`}

\item{bins:}{how many bins the histogram uses. Defaults to ggplot's default}
}
\value{
a histogram ggplot object
}
\description{
Quickly make a ggplot histogram from a single vector
}
\examples{
rnorm(1000, 0, 1) \%>\% vector_hist()


}
