% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghist.R
\name{gghist}
\alias{gghist}
\title{Quickly make a ggplot histogram}
\usage{
gghist(.data, bins = length(.data)/20)
}
\arguments{
\item{.data:}{a vector of values, such as one returned by `rnorm()`}
}
\value{
a histogram ggplot object
}
\description{
Quickly make a ggplot histogram
}
\examples{
rnorm(1000, 0, 1) \%>\% gghist()


}
